requirements.txt
berserk==0.12.4
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
ndjson==0.3.1
oauthlib==3.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
urllib3==2.0.3
